리팩토링: sdk-runtime-generator 하드코딩 제거 및 완전 동적 감지 구현
- 네임스페이스/함수 감지를 타입 주석 텍스트 기반으로 변경
- KNOWN_EVENT_NAMESPACES, KNOWN_NAMESPACE_OBJECTS, KNOWN_GLOBAL_FUNCTIONS 상수 제거
- getTypeAnnotationText()로 스택 오버플로우 없이 패턴 매칭
- addEventListener 패턴으로 이벤트 네임스페이스 자동 감지
- FunctionDeclaration 패턴으로 글로벌 함수 자동 감지
- 객체 리터럴 + 메서드 패턴으로 네임스페이스 객체 자동 감지
- deprecated 함수 자동 제외 (getSafeAreaInsets 등)
- AIT.Other.cs를 카테고리별 파일로 분리 (Advertising, IAP, Partner, SafeArea, Storage)
- collectMetaFiles 경로 해석 버그 수정 (GUID 보존)
