+++ Runtime/SDK/AIT.Advertising.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c831c564a25de9ab79d202116ae53a4a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
