+++ Runtime/SDK/AIT.IAP.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d189263571b6b8a8f111c38e43beb603
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
