+++ Runtime/SDK/AIT.Partner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3367cf20e189b0e12ca52f31a68781e1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
