+++ Runtime/SDK/AIT.SafeArea.cs
@@ -1,0 +1,71 @@
+// -----------------------------------------------------------------------
+// <copyright file="AIT.SafeArea.cs" company="Toss">
+//     Copyright (c) Toss. All rights reserved.
+//     Apps in Toss Unity SDK - SafeArea APIs
+// </copyright>
+// -----------------------------------------------------------------------
+
+using System;
+using System.Threading.Tasks;
+using Newtonsoft.Json;
+using UnityEngine.Scripting;
+
+namespace AppsInToss
+{
+    /// <summary>
+    /// Apps in Toss Platform API - SafeArea
+    /// </summary>
+    public static partial class AIT
+    {
+        /// <exception cref="AITException">Thrown when the API call fails</exception>
+        [Preserve]
+        [APICategory("SafeArea")]
+        public static async Task<SafeAreaInsetsGetResult> SafeAreaInsetsGet()
+        {
+#if UNITY_WEBGL && !UNITY_EDITOR
+            var tcs = new TaskCompletionSource<SafeAreaInsetsGetResult>();
+            string callbackId = AITCore.Instance.RegisterCallback<SafeAreaInsetsGetResult>(
+                result => tcs.TrySetResult(result),
+                error => tcs.TrySetException(error)
+            );
+            __SafeAreaInsetsGet_Internal(callbackId, "SafeAreaInsetsGetResult");
+            return await tcs.Task;
+#else
+            // Unity Editor mock implementation
+            UnityEngine.Debug.Log($"[AIT Mock] SafeAreaInsetsGet called");
+            await Task.CompletedTask;
+            return default(SafeAreaInsetsGetResult);
+#endif
+        }
+
+#if UNITY_WEBGL && !UNITY_EDITOR
+        [System.Runtime.InteropServices.DllImport("__Internal")]
+        private static extern void __SafeAreaInsetsGet_Internal(string callbackId, string typeName);
+#endif
+        /// <exception cref="AITException">Thrown when the API call fails</exception>
+        [Preserve]
+        [APICategory("SafeArea")]
+        public static async Task<System.Action> SafeAreaInsetsSubscribe(System.Action __0)
+        {
+#if UNITY_WEBGL && !UNITY_EDITOR
+            var tcs = new TaskCompletionSource<System.Action>();
+            string callbackId = AITCore.Instance.RegisterCallback<System.Action>(
+                result => tcs.TrySetResult(result),
+                error => tcs.TrySetException(error)
+            );
+            __SafeAreaInsetsSubscribe_Internal(AITJsonSettings.Serialize(__0), callbackId, "System.Action");
+            return await tcs.Task;
+#else
+            // Unity Editor mock implementation
+            UnityEngine.Debug.Log($"[AIT Mock] SafeAreaInsetsSubscribe called");
+            await Task.CompletedTask;
+            return default(System.Action);
+#endif
+        }
+
+#if UNITY_WEBGL && !UNITY_EDITOR
+        [System.Runtime.InteropServices.DllImport("__Internal")]
+        private static extern void __SafeAreaInsetsSubscribe_Internal(string __0, string callbackId, string typeName);
+#endif
+    }
+}
