+++ Runtime/SDK/AIT.SafeArea.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d6a916cb20d700b0341cd00ae6b92cf5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
